+++ src/avr/src/spindle.c
@@ -28,56 +28,90 @@
 #include "spindle.h"
 
 #include "config.h"
-#include "pwm_spindle.h"
-#include "huanyang.h"
 #include "pgmspace.h"
+
+#include <math.h>
 
 
 typedef enum {
   SPINDLE_TYPE_DISABLED,
-  SPINDLE_TYPE_PWM,
-  SPINDLE_TYPE_HUANYANG,
+#define SPINDLE(NAME) SPINDLE_TYPE_##NAME,
+#include "spindle.def"
+#undef SPINDLE
 } spindle_type_t;
+
+
+// Function decls
+#define SPINDLE(NAME)                         \
+  void NAME##_init();                         \
+  void NAME##_deinit();                       \
+  void NAME##_set(float speed);               \
+  float NAME##_get();                         \
+  void NAME##_stop();
+#include "spindle.def"
+#undef SPINDLE
 
 
 typedef struct {
   spindle_type_t type;
   float speed;
   bool reversed;
+  float min_rpm;
+  float max_rpm;
 } spindle_t;
 
 
 static spindle_t spindle = {SPINDLE_TYPE_DISABLED,};
 
 
-void spindle_init() {}
-
-
 void spindle_set_speed(float speed) {
   spindle.speed = speed;
+
+  bool negative = speed < 0;
+  if (spindle.max_rpm <= fabs(speed)) speed = negative ? -1 : 1;
+  else if (fabs(speed) < spindle.min_rpm) speed = 0;
+  else speed /= spindle.max_rpm;
 
   if (spindle.reversed) speed = -speed;
 
   switch (spindle.type) {
   case SPINDLE_TYPE_DISABLED: break;
-  case SPINDLE_TYPE_PWM: pwm_spindle_set(speed); break;
-  case SPINDLE_TYPE_HUANYANG: hy_set(speed); break;
+#define SPINDLE(NAME) case SPINDLE_TYPE_##NAME: NAME##_set(speed); break;
+#include "spindle.def"
+#undef SPINDLE
   }
 }
 
 
-float spindle_get_speed() {return spindle.speed;}
+float spindle_get_speed() {
+  float speed = 0;
+
+  switch (spindle.type) {
+  case SPINDLE_TYPE_DISABLED: break;
+#define SPINDLE(NAME) case SPINDLE_TYPE_##NAME: speed = NAME##_get(); break;
+#include "spindle.def"
+#undef SPINDLE
+  }
+
+  return speed * spindle.max_rpm;
+}
 
 
 void spindle_stop() {
   switch (spindle.type) {
   case SPINDLE_TYPE_DISABLED: break;
-  case SPINDLE_TYPE_PWM: pwm_spindle_stop(); break;
-  case SPINDLE_TYPE_HUANYANG: hy_stop(); break;
+#define SPINDLE(NAME) case SPINDLE_TYPE_##NAME: NAME##_stop(); break;
+#include "spindle.def"
+#undef SPINDLE
   }
 }
 
 
+bool spindle_is_reversed() {return spindle.reversed;}
+static void _update_speed() {spindle_set_speed(spindle.speed);}
+
+
+// Var callbacks
 uint8_t get_spindle_type() {return spindle.type;}
 
 
@@ -87,16 +121,18 @@ void set_spindle_type(uint8_t value) {
 
     switch (spindle.type) {
     case SPINDLE_TYPE_DISABLED: break;
-    case SPINDLE_TYPE_PWM: pwm_spindle_deinit(); break;
-    case SPINDLE_TYPE_HUANYANG: hy_deinit(); break;
+#define SPINDLE(NAME) case SPINDLE_TYPE_##NAME: NAME##_deinit(); break;
+#include "spindle.def"
+#undef SPINDLE
     }
 
     spindle.type = (spindle_type_t)value;
 
     switch (spindle.type) {
     case SPINDLE_TYPE_DISABLED: break;
-    case SPINDLE_TYPE_PWM: pwm_spindle_init(); break;
-    case SPINDLE_TYPE_HUANYANG: hy_init(); break;
+#define SPINDLE(NAME) case SPINDLE_TYPE_##NAME: NAME##_init(); break;
+#include "spindle.def"
+#undef SPINDLE
     }
 
     spindle_set_speed(speed);
@@ -104,7 +140,8 @@ void set_spindle_type(uint8_t value) {
 }
 
 
-bool spindle_is_reversed() {return spindle.reversed;}
+void set_speed(float speed) {spindle_set_speed(speed);}
+float get_speed() {return spindle_get_speed();}
 bool get_spin_reversed() {return spindle.reversed;}
 
 
@@ -116,6 +153,7 @@ void set_spin_reversed(bool reversed) {
 }
 
 
-// Var callbacks
-void set_speed(float speed) {spindle_set_speed(speed);}
-float get_speed() {return spindle_get_speed();}
+float get_max_spin() {return spindle.max_rpm;}
+void set_max_spin(float value) {spindle.max_rpm = value; _update_speed();}
+float get_min_spin() {return spindle.min_rpm;}
+void set_min_spin(float value) {spindle.min_rpm = value; _update_speed();}