+++ src/avr/src/usart.h
@@ -37,7 +37,7 @@
 #define USART_RX_RING_BUF_SIZE 256
 
 
-enum {
+typedef enum {
   USART_BAUD_9600,
   USART_BAUD_19200,
   USART_BAUD_38400,
@@ -48,7 +48,8 @@ enum {
   USART_BAUD_921600,
   USART_BAUD_500000,
   USART_BAUD_1000000
-};
+} baud_t;
+
 
 enum {
   USART_CRLF  = 1 << 0,
@@ -57,9 +58,37 @@ enum {
   USART_FLUSH = 1 << 3,
 };
 
+
+typedef enum {
+  USART_NONE,
+  USART_EVEN,
+  USART_ODD,
+} parity_t;
+
+
+typedef enum {
+  USART_1STOP,
+  USART_2STOP,
+} stop_t;
+
+
+typedef enum {
+  USART_5BITS,
+  USART_6BITS,
+  USART_7BITS,
+  USART_8BITS,
+  USART_9BITS,
+} bits_t;
+
+
+void usart_set_baud(USART_t *port, baud_t baud);
+void usart_set_parity(USART_t *port, parity_t parity);
+void usart_set_stop(USART_t *port, stop_t stop);
+void usart_set_bits(USART_t *port, bits_t bits);
+void usart_init_port(USART_t *port, baud_t baud, parity_t parity, bits_t bits,
+                     stop_t stop);
+
 void usart_init();
-void usart_set_port_baud(USART_t *port, int baud);
-void usart_set_baud(int baud);
 void usart_set(int flag, bool enable);
 bool usart_is_set(int flags);
 void usart_putc(char c);